requirements.txt
altgraph==0.17.2
attrs==20.3.0
bidict==0.21.3
funcy==1.16
macholib==1.15.2
parsimonious==0.8.1
py-aiger==6.1.14
py-aiger-cnf==5.0.2
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pypblib==0.0.4
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.17.3
python-sat==0.1.7.dev10
PyYAML==6.0
six==1.16.0
sortedcontainers==2.4.0
toposort==1.6
